// src/client/types.gen.ts
export type ShareDeletePageShareData = {
	body?: never
	headers: {
		/**
		 * Required to be true for the API request to pass
		 */
		'OCS-APIRequest': boolean
	}
	path: {
		apiVersion: '1.0'
		/**
		 * ID of the collective
		 */
		collectiveId: number
		/**
		 * ID of the page
		 */
		pageId: number
		/**
		 * Token of the share
		 */
		token: string
	}
	query?: never
	url: '/ocs/v2.php/apps/collectives/api/v{apiVersion}/collectives/{collectiveId}/pages/{pageId}/shares/{token}'
}
// ...
export type ShareUpdatePageShareData = {
	body: {
		/**
		 * Whether share has edit permissions
		 */
		editable: boolean
		/**
		 * Optional password for the share
		 */
		password?: string | null
	}
	headers: {
		/**
		 * Required to be true for the API request to pass
		 */
		'OCS-APIRequest': boolean
	}
	path: {
		apiVersion: '1.0'
		/**
		 * ID of the collective
		 */
		collectiveId: number
		/**
		 * ID of the page
		 */
		pageId: number
		/**
		 * Token of the share
		 */
		token: string
	}
	query?: never
	url: '/ocs/v2.php/apps/collectives/api/v{apiVersion}/collectives/{collectiveId}/pages/{pageId}/shares/{token}'
}